Implement basic functionality
This performs basic password password decoding,
serialization, and substitutes one character at
a time to find potential errors.

There is no attempt to sort errors and find the
most likely one based on character or phonetic
similarity, or data correctness, so all possible
corrections are written out for the user.